refactor(fp8): lift the row-wise family out of dispatch_proj.rs
dispatch_proj.rs was 445 LoC before this branch and 538 after — over the
repo's 500 cap, and unlike the six violations inherited from #516 this one
is ours. CI was red on it.

Split by cohesion rather than by line count: the passthrough decision, the
cached block->row-wise requant it guards, the router that consumes both,
and their CPU-only tests are one unit and move together. 538 -> 314, with a
245-line sibling. dispatch_proj.rs is off the violation list; the six that
remain are all pre-existing.

The new file needs no `use super::*` because everything in it names its
paths explicitly — worth saying, because dispatch_proj.rs carries
`#![allow(unused_imports)]` and would have hidden an unused glob.

Behaviour-identical: same 542 spark-model tests pass, and the two
passthrough tests just move module path (dispatch_proj -> the sibling).
fmt/clippy/typos clean.
